+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log (@egomobile/queue)
 
+## 2.0.0
+
+- BREAKING CHANGE: rename [queueTask](https://egomobile.github.io/node-queue/interfaces/IQueueStorage.html#queueTask) to `enqueueTask`
+- add `enqueueRemaing()` method to [IQueueStorage](https://egomobile.github.io/node-queue/interfaces/IQueueStorage.html)
+- fix [MemoryQueueStorage](https://egomobile.github.io/node-queue/classes/MemoryQueueStorage.html)
+
 ## 1.0.1
 
 - initial release